% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_standard package.R
\name{country_standard}
\alias{country_standard}
\title{country_standard}
\usage{
country_standard(x = NULL, code = "ISOA3", name = "ISOname")
}
\arguments{
\item{x}{A character vector of country names, either spelled correctly or with misspellings, that you wish to standardize}

\item{code}{A character vector of country code. Defaults to "ISOA3", options: "ISOA3", "ISOA2", "ISON", "IMFcode", "FIPS", "STANAG", "Internet"}

\item{name}{A character vector of country name. Defaults to "ISOname", options: "ISOname", "IMFname", "FIPSname"}
}
\value{
a character vector of supplied country names, a vector of corrected country names, and and optional vector of country codes
}
\description{
country_standard
}
\examples{
country_standard(c("congo", "congo dr", "Democratic People's Republic of Congo"), code="IMFcode", name="IMFname")
}

